aula 12 finalizando curso

+++ aula06/gerenciador_de_produtos/src/pages/Product/ProductList.jsx
@@ -1,67 +1,68 @@
-/* eslint-disable no-unused-vars */
-
-import React, { useEffect } from 'react'
-import { useState } from 'react'
+import React from 'react'
 import { useNavigate } from 'react-router-dom'
+import { useState, useEffect } from 'react'
 import axios from '../../api'
 
 const ProductList = () => {
+
     const [products, setProducts] = useState([])
-    const navigate = useNavigate
+    const navigate = useNavigate()
 
     useEffect(() => {
-        axios.get('/products?_expand=supplier')
-            .then(response => {
-                setProducts(response.data)
-            })
-            .catch(error => console.error("There was an error!", error))
+        axios.get('/products')
+        .then(response => {
+            setProducts(response.data)
+        })
+        .catch(error => console.error("Ocorreu um erro: ",error))
     }, []);
 
     const fetchProducts = () => {
-        axios.get('/products?_expand=supplier')
+        axios.get('/products')
             .then(response => {
                 setProducts(response.data)
             })
-            .catch(error => console.error("There was an error!", error))
+            .catch(error => console.error("Ocorreu um erro: ",error))
         }
+    
+    function deleteProduct(id){
 
-        function deleteProduct(id) {
-            axios.delete(`/products/${id}`)
-                .then(() => {
-                    alert("Produto excluido com sucesso!")
-                    fetchProducts()
-        
-                })
-                .catch(error => console.error("There was an error!", error))
-        }
+        if(confirm("Tem certeza que voce quer deletar?")){
+        axios.delete(`/products/${id}`)
+        .then(() => {
+            alert("Produto excluído com sucesso!")
+            fetchProducts()
+        })
+        .catch(error => console.error("Ocorreu um erro: ",error))
+    }
+     }
+      
+
   return (
-    <div className='container mt-5'>
-        <h2 className='mb-4'>Lista de Produtos</h2>
-        <button onClick={() => navigate('/add-produto')} className='btn btn-primary mb-2'>Adicionar Produto</button>
-        <table className='table'>
+    <div className="container mt-5">
+        <h2 className="mb-4">Lista de Produtos</h2>
+        <button onClick={() => navigate('/add-produto')} className="btn btn-primary mb-2">Adicionar Produto</button>
+        <table className="table">
             <thead>
                 <tr>
-                    <th>Nome</th>
-                    <th>Preço</th>
-                    <th>Fornecedor</th>
-                    <th>Ações</th>
+                    <th>Nome:</th>
+                    <th>Preço:</th>
+                    <th>Fornecedor:</th>
+                    <th>Ações:</th>
                 </tr>
             </thead>
             <tbody>
                 {
-                    products.map(product => (
-                        <tr key={product.id}>
-                            <td>{product.name}</td>
-                            <td>{product.price}</td>
-                            <td>{product.supplier.name}</td>
-                            <td>
-                                <button className='btn btn-sm btn-warning mr-2'onClick={() => navigate(`/editar-produto/${product.id}`)}>Editar</button>
-                                <button onClick={() => deleteProduct(product.id)} className='btn btn-sm btn-danger mr-2'>Excuir</button>
-
-                            </td>
-
-                        </tr>
-                    ))
+                   products.map(product => (
+                    <tr key={product.id}>
+                        <td>{product.name}</td>
+                        <td>R${Number(product.price).toFixed(2)}</td>
+                        <td>{product.productSupplierDetails.name}</td> 
+                        <td>
+                            <button className="btn btn-sm btn-warning mr-2" onClick={() => navigate(`/editar-produto/${product.id}`)}>Editar</button>
+                            <button onClick={() => deleteProduct(product.id)} className="btn btn-sm btn-danger">Excluir</button>
+                        </td>        
+                    </tr>
+                ))
                 }
             </tbody>
         </table>
@@ -70,4 +71,4 @@ const ProductList = () => {
   )
 }
 
-export default ProductList
+export default ProductList
